requirements.txt
art==3.6
capstone==4.0.1
certifi==2019.6.16
Click==7.0
colorama==0.4.1
coverage==4.5.3
future==0.17.1
joblib==0.13.2
numpy==1.16.4
pandas==0.24.2
pefile==2019.4.18
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tqdm==4.32.2
wincertstore==0.2
xgboost==0.90
yara-python==3.10.0
